refactor: separate backend API endpoints into individual view files
- Split monolithic views.py (1,493 lines) into focused view files
- Created views package with 7 dedicated files:
  * code_layout_views.py - CodeLayoutViewSet
  * page_theme_views.py - PageThemeViewSet
  * widget_type_views.py - WidgetTypeViewSet
  * webpage_views.py - WebPageViewSet
  * page_version_views.py - PageVersionViewSet
  * page_data_schema_views.py - PageDataSchemaViewSet
  * rendering_views.py - Function-based rendering views
- Maintained full backward compatibility via __init__.py
- Fixed smartSaveUtils.js version ID handling
- Removed debug statements
